pack: Break out Session type to a seperate module
